+++ lib/rubocop/cop/rspec/indexed_let.rb
@@ -31,12 +31,18 @@ module RuboCop
       #   let(:item_1) { create(:item) }
       #   let(:item_2) { create(:item) }
       #
+      # @example `AllowedIdentifiers: ['item_1', 'item_2']`
+      #   # good
+      #   let(:item_1) { create(:item) }
+      #   let(:item_2) { create(:item) }
+      #
       # @example `AllowedPatterns: ['item']`
       #   # good
       #   let(:item_1) { create(:item) }
       #   let(:item_2) { create(:item) }
       #
       class IndexedLet < Base
+        include AllowedIdentifiers
         include AllowedPattern
 
         MSG = 'This `let` statement uses index in its name. Please give it ' \
@@ -78,6 +84,7 @@ module RuboCop
         def indexed_let?(node)
           let?(node) &&
             SUFFIX_INDEX_REGEX.match?(let_name(node)) &&
+            !allowed_identifier?(let_name(node).to_s) &&
             !matches_allowed_pattern?(let_name(node).to_s)
         end
 
@@ -89,6 +96,12 @@ module RuboCop
           Array(config.for_cop('Naming/VariableNumber')
             .fetch('AllowedPatterns', [])) +
             Array(cop_config.fetch('AllowedPatterns', []))
+        end
+
+        def allowed_identifiers
+          Array(config.for_cop('Naming/VariableNumber')
+            .fetch('AllowedIdentifiers', [])) +
+            Array(cop_config.fetch('AllowedIdentifiers', []))
         end
       end
     end