+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update the message output of `RSpec/ExpectActual` to include the word 'value'. ([@corydiamand])
 - Fix a false negative for `RSpec/Pending` when  `it` without body. ([@ydah])
 - Add new `RSpec/ReceiveMessages` cop. ([@ydah])
+- Change default.yml path to use `**/spec/*` instead of `spec/*`. ([@ydah])
 
 ## 2.22.0 (2023-05-06)