+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add new `RSpec/ReceiveMessages` cop. ([@ydah])
 - Add `AllowedIdentifiers` and `AllowedPatterns` configuration option to `RSpec/IndexedLet`.  ([@ydah])
 - Fix `RSpec/NamedSubject` when block has no body. ([@splattael])
+- Fix `RSpec/LetBeforeExamples` autocorrect incompatible with `RSpec/ScatteredLet` autocorrect. ([@ydah])
 
 ## 2.22.0 (2023-05-06)